Issue #128 Make JsonPath an interface
